+++ pathlib3x/__init__conf__.py
@@ -1,6 +1,6 @@
 name = 'pathlib3x'
-title = 'backport of latest pathlib to python 3.6, 3.7 with a few extensions'
-version = '0.1.1'
+title = 'backport of pathlib 3.10 to python 3.6, 3.7, 3.8, 3.9 with a few extensions'
+version = '0.2.0'
 url = 'https://github.com/bitranox/pathlib3x'
 author = 'Robert Nowotny'
 author_email = 'bitranox@gmail.com'
@@ -12,9 +12,9 @@ def print_info() -> None:
 
 Info for pathlib3x:
 
-    backport of latest pathlib to python 3.6, 3.7 with a few extensions
+    backport of pathlib 3.10 to python 3.6, 3.7, 3.8, 3.9 with a few extensions
 
-    Version : 0.1.1
+    Version : 0.2.0
     Url     : https://github.com/bitranox/pathlib3x
     Author  : Robert Nowotny
     Email   : bitranox@gmail.com""")