+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog
 - new MINOR version for added functionality in a backwards compatible manner
 - new PATCH version for backwards compatible bug fixes
 
+0.2.0
+-----
+2020-07-02 : added function: PurePath.append_suffix(suffix)
+    - added function: PurePath.append_suffix(suffix)
+
 0.1.1
 -----
 2020-07-01: maintainance release